% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidatePack-package.R
\docType{package}
\name{candidatePack}
\alias{candidatePack}
\alias{candidatePack-package}
\title{candidatePack}
\description{
The candidatePack package looks at presidential candidate delegates and performs calculations
}
\examples{

\dontrun{
demo(addSquares)
demo(subtractSquares)
}

}
\author{
Scott Solomon: \email{scott.solomon@wustl.edu} and
}

