% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_candidate.R
\name{create_candidate}
\alias{create_candidate}
\title{taking objects and putting them into class "Candidate"}
\usage{
create_candidate(name, delegates_won, party)
}
\arguments{
\item{name}{a character object}

\item{party}{a character object}

\item{delegatesWon}{a numeric object}
}
\value{
An object of class Candidate containing
 \item{names}{The name of the candidate}
 \item{delegates_won}{The number of delegates won} 
 \item{party}{The party of the candidate}
 \item{delegates_needed}{the number of delegates needed to win the primary}
}
\description{
takes objects and puts them into the proper class slots
}
\note{
Make america gr8 again
}
\examples{

create_candidate(name = "Obama", delegates_won = 800, party = "Democrat")
}
\author{
Scott Solomon
}

