% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_candidate.R
\docType{methods}
\name{print,Candidate-method}
\alias{print,Candidate-method}
\title{taking an object of class "Candidate" and using the generic print function}
\usage{
\S4method{print}{Candidate}(x)
}
\arguments{
\item{name}{a character object}

\item{delegates_won}{a numeric object}

\item{party}{a character object}
}
\value{
the output of print for the "Candidate" object
 \item{names}{The name of the candidate}
 \item{delegates_won}{The number of delegates won} 
 \item{party}{The party of the candidate}
 \item{delegates_needed}{the number of delegates needed to win the primary}
}
\description{
takes an object of class "Candidate" and executes print
}
\note{
Make america gr8 again
}
\examples{

Obama<-new("Candidate", name = "Obama", delegates_won = 800, party = "Democrat", delegates_needed = 1200)
print(Obama)
}
\author{
Scott Solomon
}

