% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cand.R
\name{}
\alias{print_cand}
\alias{print_cand,ANY-method}
\title{Prints the candidate and his/her info}
\usage{
Candidate(name, party, delegatesNeeded, delegatesWon)
}
\arguments{
\item{name}{The name of the candidate running}

\item{party} Which political party is the candidate representing.

\item{delegatesWon} How many delegates the candidate has won so far.

\item{delegatesNeeded} How many delegates the candidate needs to win the nomination.
}
\value{
An object of class Candidate containing
\item{name} The name of the candidate running.
\item{party} Which political party is the candidate representing
\item{delegatesWon} How many delegates the candidate has won so far.
\item{delegatesNeeded} How many delegates the candidate needs to win the nomination.
}
\description{
Finds the sum of squared numbers
}
\author{
Joseph Ludmir
}


