% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_needed.R
\name{prop_needed}
\alias{prop_needed}
\title{taking an object of class "Candidate" and gives the percent of delegates left needed to win to take the nomination}
\usage{
prop_needed(candidate, remaining_delegates)
}
\arguments{
\item{Candidate}{a "candidate" object}
}
\value{
the output of print for the "Candidate" object
 \item{percent_needed}{The percent of remaining delegates needed}
}
\description{
takes an object of class "Candidate" and executes a calculation to see what percent of delegates are needed
}
\note{
Make america gr8 again
}
\examples{

Obama<-new("Candidate", name = "Obama", delegates_won = 800, party = "Democrat", delegates_needed = 1200)
prop_needed(Obama)
}
\author{
Scott Solomon
}

