% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded_cand.R
\name{propNeeded_cand}
\alias{propNeeded_cand}
\alias{propNeeded_cand,ANY-method}
\title{ Calculate percentage of remaining delegates needed by a specific presidential candidate.}
\usage{
Candidate(name, party, delegatesNeeded, delegatesWon)
}
\arguments{
\item{name}{The name of the candidate running}

\item{party} Which political party is the candidate representing.

\item{delegatesWon} How many delegates the candidate has won so far.

\item{delegatesNeeded} How many delegates the candidate needs to win the nomination.
}
\value{
An object of class Candidate containing
\item{name} The name of the candidate running.
\item{party} Which political party is the candidate representing
\item{delegatesWon} How many delegates the candidate has won so far.
\item{delegatesNeeded} How many delegates the candidate needs to win the nomination.
}
\author{
Joseph Ludmir
}


