% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race.R
\docType{class}
\name{Race-class}
\alias{Race}
\alias{Race-class}
\alias{initialize,Race-method}
\alias{plot,Race-method}
\title{An object of class Race}
\description{
Objects of class \code{Race} summarize the standing of the different candidates in a given primary race.
}
\details{
An object of the class `Race' has the following slots:
\itemize{
\item \code{candidates} A list of objects of class `Candidate', of the same party.
\item \code{party} The party of the presidential candidates listed.
\item \code{remainingDelegates} The number of delegates left. 
}
}
\examples{
kasich <- new("Candidate", name="Kasich", delegatesWon=63, party="Rep")
miguel <- new("Candidate", name="Miguel", delegatesWon=35, party="Rep")
republican.candidates <- list(kasich, miguel)
republican.race <- new("Race", candidates=republican.candidates, party="Rep")
plot(x=republican.race)

}
\author{
Miguel Pereira: \email{m.pereira@wustl.edu}
}
\seealso{
\code{\link{createCandidate}}, \code{\link{Candidate}}
}

