% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\name{Candidate}
\alias{Candidate}
\alias{Candidate,ANY-method}
\title{A candidate object.}
\usage{
Candidate(name, party, delegatesNeeded, delegatesWon)
}
\arguments{
\item{name}{The name of the candidate running}

\item{party} Which political party is the candidate representing.

\item{delegatesWon} How many delegates the candidate has won so far.

\item{delegatesNeeded} How many delegates the candidate needs to win the nomination.
}
\value{
An object of class Candidate containing
\item{name} The name of the candidate running.
\item{party} Which political party is the candidate representing
\item{delegatesWon} How many delegates the candidate has won so far.
\item{delegatesNeeded} How many delegates the candidate needs to win the nomination.
}
\author{
Joseph Ludmir
}


