% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\alias{Candidate-method}
\alias{initialize,}
\title{An of candidate information}
\description{
Object of class \code{delegatePack} take information about a candidates delegates
}
\details{
An item of class `Candidate' has the following slots:
\itemize
\item \code{name} the name of the candidate
\item \code{delegates_won} the number of delegates the candidate has won
\item \code{party} which party the delegate is in
\item \code{delegates_needed} the number of delegates still needed, determined in create_candidate
}
\author{
Scott Solomon \email{scott.solomon@wustl.edu}
}

