% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\alias{initialize,Candidate-method}
\alias{show,Candidate-method}
\title{A Candidate object}
\description{
Object of class \code{Candidate} are created by the \code{createCandidate} function. Objects of
this class have their own \code{show} and \code{print} methods.
}
\details{
An object of the class `Candidate' has the following slots:
\itemize{
\item \code{name} The name of a presidential candidate.
\item \code{delegatesWon} The number of delegates the presidential candidate has won.
\item \code{party} The party of the presidential candidate, either `Democratic' or `Republican'.
\item \code{delegatesNeeded} The number of delegates the presidential candidate needs to secure to win his party's nomination.
}
}
\author{
Andy Stone: \email{arstone@wustl.edu}
}

