% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidatePack-package.r
\docType{package}
\name{candidatePack}
\alias{candidatePack}
\alias{candidatePack-package}
\title{candidatePack}
\description{
The candidatePack package allows the user to examine delegate counts for the 2016 presidential election.
}
\examples{

\dontrun{
demo(createCandidate)
demo(propNeeded)
}

}
\author{
Andy Stone: \email{arstone@wustl.edu}
}

