% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCandidate.R
\name{createCandidate}
\alias{createCandidate}
\alias{createCandidate,ANY-method}
\title{Creating an object of class `Candidate'}
\usage{
createCandidate(name, delegatesWon, party)
}
\arguments{
\item{name}{A character string giving the name of the presidential candidate.}

\item{delegatesWon}{A numeric object listing how many delegates the candidate has already won.}

\item{party}{A character string denoting the party of the candidate. This string should be given
as either `Republican' or `Democratic'.}
}
\value{
An object of class `Candidate' containing
 \item{name} The name of a presidential candidate.
 \item{delegatesWon} The number of delegates the presidential candidate has won. 
 \item{party} The party of the presidential candidate, either `Democratic' or `Republican'.
 \item{delegatesNeeded} The number of delegates the presidential candidate needs to secure to win 
 his party's nomination. This is automatically determined by calling the neededCalculator() function
 and passing it the party parameter.
}
\description{
A user-friendly function to create an object of class `Candidate' and automatically determine
how many delegates the candidate needs to secure their party's presidential nomination.
}
\note{
delegatesWon can represent all delegates or just pledged delegates -- but keep this in mind.
}
\examples{

createCandidate("Jim Webb", 0, "Democratic") 
createCandidate("Marco Rubio", 155, "Republican") 

}
\author{
Andy Stone: \email{arstone@wustl.edu}
}
\seealso{
\code{\link{neededCalculator}}
}

