% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded.R
\name{propNeeded}
\alias{propNeeded}
\alias{propNeeded,ANY-method}
\title{Calculating proportion of candidates needed to win nomination}
\usage{
propNeeded(candidate, remaining)
}
\arguments{
\item{candidate}{An object of the class `Candidate'}

\item{remaining}{A numeric object denoting the number of delegates yet to be allocated in the 
party of the candidate. This should include both pledged and unpledged delegates.}
}
\value{
A numeric object denoting the proportion of the delegates yet to be allocated that a 
presidential candidate must secure in order to win his party's nomination.
}
\description{
A function to determine the proportion of remaining delegates a presidential candidate needs 
to secure in order to win his party's nomination.
}
\note{
Be certain to include both pledged and unpledged delegates in the total given for
\code{remaining}.
}
\examples{

candidate.rubio <- createCandidate("Marco Rubio", 155, "Republican")
propNeeded(candidate.rubio, 1407) 

}
\author{
Andy Stone: \email{arstone@wustl.edu}
}
\seealso{
\code{\link{createCandidate}}, \code{\link{Candidate}}
}

