% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race.R
\docType{class}
\name{Race-class}
\alias{Race}
\alias{Race-class}
\alias{initialize,Race-method}
\alias{plot,Race-method}
\title{A Race object}
\description{
Obects of class \code{Race} help summarize the standing of the current presidential races. These
objects have their own \code{print} methods.
}
\details{
An object of the class `Race' has the following slots:
\itemize{
\item \code{candidates} A list of objects of class `Candidate'. These candidates should all 
be from the same political party.
\item \code{party} The party of the presidential candidates listed in candidates, either 
`Democratic' or `Republican'.
\item \code{delegatesNeeded} The number of delegates the presidential candidate needs to 
secure to win his party's nomination.
\item \code{delegatesRemaining} The number of delegates that have yet to be allocated to a 
candidate. 
}
}
\author{
Andy Stone: \email{arstone@wustl.edu}
}

