% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race.R
\docType{class}
\name{Race-class}
\alias{Race}
\alias{Race-class}
\alias{initialize,Race-method}
\alias{plot,Race-method}
\title{A Race object}
\description{
Obects of class \code{Race} help summarize the standing of the current presidential races. These
objects have their own \code{plot} method.
}
\details{
An object of the class `Race' has the following slots:
\itemize{
\item \code{candidates} A list of objects of class `Candidate'. These candidates should all 
be from the same political party.
\item \code{party} The party of the presidential candidates listed in candidates, either 
`Democratic' or `Republican'.
\item \code{delegatesNeeded} The number of delegates the presidential candidate needs to 
secure to win his party's nomination.
\item \code{delegatesRemaining} The number of delegates that have yet to be allocated to a 
candidate. 
}
}
\examples{
cruz <- new("Candidate", name="Cruz", delegatesWon=372, party="Republican")
trump <- new("Candidate", name="Trump", delegatesWon=464, party="Republican")
kasich <- new("Candidate", name="Kasich", delegatesWon=63, party="Republican")
rubio <- new("Candidate", name="Rubio", delegatesWon=166, party="Republican")
republican.candidates <- list(cruz, trump, kasich, rubio)
republican.race <- new("Race", candidates=republican.candidates, party="Republican")
plot(x=republican.race)

}
\author{
Andy Stone: \email{arstone@wustl.edu}
}
\seealso{
\code{\link{createCandidate}}, \code{\link{Candidate}}
}

